Some cleaning and clarifying
