+++ Gameplan/Game/Checkers/CheckersPlayer.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Player.h"
+#include "../Player.h"
 #include <iostream>
 
 struct CheckersPlayer : public Player
