+++ Gameplan/Game/Checkers/CheckersTable.cpp
@@ -1,4 +1,4 @@
-#include "../application.h"
+#include "../../application.h"
 #include "CheckersTable.h"
 #include "CheckersPlayer.h"
 
@@ -8,19 +8,20 @@ void CheckersTable::update(class Player* _player, class Application* _applicatio
 	{
 		if (_application->mouse.clicked_left)
 		{
+
+			// Get clicked slot's column and row
 			int column = board.getColumn(_application->mouse.pos.x);
 			int row = board.getRow(_application->mouse.pos.y);
 
-			if (row > -1 && column > -1)
+			if (board.isSlotValid({row, column}))
 			{
-				// Selected square has pawn
-				if (board.slots[row][column].hasPawn)
+				// Selected square has pawn -> select it
+				if (board.slots[row][column].hasPawn) // board.isPawnInSlot({row, column})
 				{
-					selected_slot = board.slots[row][column];
-					selected_slot.pawn.board_slot = { column, row };
-					// can player eat -> force to eat
+					selected_slot = board.slots[row][column]; // board.getSlot({row, column})
+
 					bool can_palyer_eat = canPlayerEat(dynamic_cast<CheckersPlayer*>(_player));
-					//std::cout << can_palyer_eat << std::endl;
+
 					board.getDiagonalMoves(selected_slot.pawn, dynamic_cast<CheckersPlayer*>(_player), can_palyer_eat);
 				}
 				else if (selected_slot.hasPawn) // Already selected pawn
@@ -52,4 +53,20 @@ void CheckersTable::drawTable(class Application* _application)
 	board.DrawBoard(_application);
 	board.DrawPawns(_application);
 	board.DrawPossibleMoves(_application, selected_slot);
+}
+
+bool CheckersTable::canPlayerEat(CheckersPlayer* player)
+{
+	for (int r = 0; r < ROWS; r++)
+	{
+		for (int c = 0; c < COLUMNS; c++)
+		{
+			if (board.slots[r][c].hasPawn && board.slots[r][c].pawn.color == player->color)
+			{
+				if (board.canPawnEat(board.slots[r][c].pawn, player))
+					return true;
+			}
+		}
+	}
+	return false;
 }
