+++ Gameplan/Gameplan.vcxproj
@@ -152,9 +152,11 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="application.h" />
-    <ClInclude Include="Game\Checkers.h" />
-    <ClInclude Include="Game\CheckersPlayer.h" />
-    <ClInclude Include="Game\CheckersTable.h" />
+    <ClInclude Include="Game\Checkers\CheckersBoard.h" />
+    <ClInclude Include="Game\Checkers\CheckersBoardSlot.h" />
+    <ClInclude Include="Game\Checkers\CheckersPawn.h" />
+    <ClInclude Include="Game\Checkers\CheckersPlayer.h" />
+    <ClInclude Include="Game\Checkers\CheckersTable.h" />
     <ClInclude Include="Game\ConnectFourPlayer.h" />
     <ClInclude Include="Game\ConnectFourTable.h" />
     <ClInclude Include="Game\Game.h" />
@@ -163,6 +165,7 @@
     <ClInclude Include="Networking\Client.h" />
     <ClInclude Include="Networking\Server.h" />
     <ClInclude Include="Networking\ZipPacket.h" />
+    <ClInclude Include="resource.h" />
     <ClInclude Include="UI\Button.h" />
     <ClInclude Include="UI\UI.h" />
     <ClInclude Include="UI\UiInput.h" />
@@ -171,8 +174,8 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="application.cpp" />
-    <ClCompile Include="Game\Checkers.cpp" />
-    <ClCompile Include="Game\CheckersTable.cpp" />
+    <ClCompile Include="Game\Checkers\CheckersBoard.cpp" />
+    <ClCompile Include="Game\Checkers\CheckersTable.cpp" />
     <ClCompile Include="Game\ConnectFourPlayer.cpp" />
     <ClCompile Include="Game\ConnectFourTable.cpp" />
     <ClCompile Include="Game\GameTable.cpp" />
@@ -187,6 +190,9 @@
     <ClCompile Include="UI\UiInput.cpp" />
     <ClCompile Include="UI\UiSlider.cpp" />
     <ClCompile Include="UI\UiTextInput.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ResourceCompile Include="Gameplan.rc" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
